rename _parser.js => _parser_mock.js so it's obvious what class of module it is
